Added a simple shell program.
This is the first commit for a simple shell program.
As of right now most of the code is taken from Anders H
who developed this and provided the code at
https://gist.github.com/parse/966049. Many additions
still need to be done and the code should end up being
very different in the end. Support for pipes and I/O
redirection is forthcoming.