Some bugs fixed for I/O redirect
The arguments for execcp were incorrect.  Not only the < or >
need to be taken out of the arguments but also the argument after
them since that is already being saved as the file name.  This
fix introduced another bug - a segmentation fault for all I/O
redirect attempts.  The for loop needs to be examined more closely
to identify the issue.  Additionally the < functionality remains
undone.